--- src/errors/types.ts
@@ -1,11 +1,0 @@
-export class HttpError extends Error {
-  public status: number
-  public message: string
-
-  constructor(status: number, message: string) {
-    super(message)
-
-    this.status = status
-    this.message = message
-  }
-}
